Generalize common combat state props, add more debug logging for common event() exception
