+++ services/app/src/components/views/quest/cardtemplates/combat/DefeatContainer.test.tsx
@@ -23,6 +23,7 @@ function setup(overrides: Partial<Props>) {
     settings: initialSettings,
     combat: newCombat(TEST_NODE),
     node: TEST_NODE.clone(),
+    seed: '123',
     mostRecentRolls: undefined,
     onCustomEnd: jasmine.createSpy('onCustomEnd'),
     onEvent: jasmine.createSpy('onEvent'),
