+++ services/app/src/components/views/quest/cardtemplates/combat/DrawEnemies.tsx
@@ -3,17 +3,15 @@ import Button from 'app/components/base/Button';
 import Card from 'app/components/base/Card';
 import Picker from 'app/components/base/Picker';
 import {Enemy} from 'app/reducers/QuestTypes';
-import {SettingsType} from 'app/reducers/StateTypes';
 import * as React from 'react';
 import {REGEX} from 'shared/Regex';
 import {NUMERALS} from '../Render';
 import {ParserNode} from '../TemplateTypes';
 import {CombatPhase, CombatState} from './Types';
+import {StateProps as StatePropsBase} from './Types';
 
-export interface StateProps {
-  node: ParserNode;
+export interface StateProps extends StatePropsBase {
   combat: CombatState;
-  settings: SettingsType;
   tier: number;
 }
