+++ services/app/src/components/views/quest/cardtemplates/combat/PrepareTimerContainer.tsx
@@ -1,17 +1,10 @@
-import {AppStateWithHistory} from 'app/reducers/StateTypes';
 import {connect} from 'react-redux';
 import Redux from 'redux';
 import {
   handleCombatTimerStart,
 } from './Actions';
-import PrepareTimer, {DispatchProps, StateProps} from './PrepareTimer';
-
-const mapStateToProps = (state: AppStateWithHistory, ownProps: Partial<StateProps>): StateProps => {
-  return {
-    node: ownProps.node || state.quest.node,
-    settings: state.settings,
-  };
-};
+import PrepareTimer, {DispatchProps} from './PrepareTimer';
+import {mapStateToProps} from './Types';
 
 const mapDispatchToProps = (dispatch: Redux.Dispatch<any>): DispatchProps => {
   return {
