+++ services/app/src/components/views/quest/cardtemplates/combat/Types.test.tsx
@@ -1,3 +1,5 @@
 describe('Combat template types', () => {
-  // Entirely glue code; no testing needed right now.
+  describe('mapStateToProps', () => {
+    it('handles unset ownProps.node');
+  });
 });
